+++ lib/ts/recipe/thirdparty/components/themes/signInAndUpCallback/index.tsx
@@ -19,7 +19,6 @@ import { PureComponent } from "react";
 
 import SpinnerIcon from "../../../../../components/assets/spinnerIcon";
 import { withOverride } from "../../../../../components/componentOverride/withOverride";
-import { hasFontDefined } from "../../../../../styles/styles";
 import SuperTokens from "../../../../../superTokens";
 import { ThemeBase } from "../themeBase";
 
@@ -53,12 +52,9 @@ const SignInAndUpCallbackThemeWithOverride = withOverride(
 
 export const SignInAndUpCallbackTheme = (props: { config: NormalisedConfig }) => {
     const rootStyle = SuperTokens.getInstanceOrThrow().rootStyle;
-    const hasFont = hasFontDefined(rootStyle) || hasFontDefined(props.config.recipeRootStyle);
 
     return (
-        <ThemeBase
-            loadDefaultFont={!hasFont}
-            userStyles={[rootStyle, props.config.recipeRootStyle, props.config.signInAndUpFeature.style]}>
+        <ThemeBase userStyles={[rootStyle, props.config.recipeRootStyle, props.config.signInAndUpFeature.style]}>
             <SignInAndUpCallbackThemeWithOverride />
         </ThemeBase>
     );