+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.45.0] - 2024-07-31
+
+### Breaking changes
+
+-   Updated the font and font-weights in the default styles
+-   We no longer load the Rubik font by default
+
 ## [0.44.0] - 2024-07-15
 
 -   Removes the default `maxAgeInSeconds` value (previously 300 seconds) in EmailVerification Claim. If the claim value is true and `maxAgeInSeconds` is not provided, it will not be refreshed.